+++ actions/microsoft-mail/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.0] - 2024-10-18
+
+### Add
+
+- Action `flag_email` added.
+- Add parameter `max_emails_to_return` to action `list_emails` to minimize action response size
+
 ## [1.0.4] - 2024-10-03
 
 ### Changed